% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{AnnotateCcubeResults}
\alias{AnnotateCcubeResults}
\title{Annotate mydata with Ccube results}
\usage{
AnnotateCcubeResults(ssm, res)
}
\arguments{
\item{ssm}{Ccube input data}

\item{res}{Ccube results list}
}
\value{
ssm Annotated input data
}
\description{
Annotate mydata with Ccube results
}
