% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{AnnotateCcubeResults_sv}
\alias{AnnotateCcubeResults_sv}
\title{Annotate mydata with CcubeSV results}
\usage{
AnnotateCcubeResults_sv(ssm, res)
}
\arguments{
\item{ssm}{CcubeSV input data}

\item{res}{CcubeSV results list}
}
\value{
ssm Annotated input data
}
\description{
Annotate mydata with CcubeSV results
}
