% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Assign}
\alias{Assign}
\title{Assign data to centers}
\usage{
Assign(x, centers, s)
}
\arguments{
\item{x}{datas}

\item{centers}{cluster centers}

\item{s}{cluster variances}
}
\value{
assignment probabilites and log probabilites
}
\description{
Assign data to centers
}

