% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{AssignWithCcube}
\alias{AssignWithCcube}
\title{Computing responsibilities (assignment probabilities) with modified Ccube VBEM-step}
\usage{
AssignWithCcube(res, ssm, tol = 1e-08, maxiter = 100, epi = 0.001,
  verbose = F)
}
\arguments{
\item{res}{A reference Ccube results list}

\item{ssm}{A data frame of SNVs to be assigned. Ideally, the data has been processed by Ccube model. So it should have the ccube_mult column.}

\item{tol}{convergence threshold}

\item{maxiter}{maximum iterations, default number is 100.}

\item{epi}{sequencing error, default is 1e-3}

\item{verbose}{show VBEM progress}
}
\value{
a standard Ccube model with recomputed responsibilities and logResponsibilities
}
\description{
Computing responsibilities (assignment probabilities) with modified Ccube VBEM-step
}
