% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bsxfun.se}
\alias{bsxfun.se}
\title{bsxfun {pracma} with single expansion (real Matlab style)}
\usage{
bsxfun.se(func, x, y, expandByRow = TRUE)
}
\arguments{
\item{func}{the function used by bsxfun}

\item{x}{a matrix}

\item{y}{a vector need to be expanded}

\item{expandByRow}{applies only when x is a square matrix}
}
\value{
value of func
}
\description{
bsxfun {pracma} with single expansion (real Matlab style)
}
