% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{ccube_m6}
\alias{ccube_m6}
\title{Run Ccube with model 6: Normal-Binomial}
\usage{
ccube_m6(mydata, epi = 0.001, init = 2, prior = NULL, tol = 1e-20,
  maxiter = 1000, fit_mult = F, fit_hyper = T, use = c("use_base",
  "use_one"), verbose = FALSE)
}
\arguments{
\item{mydata}{mutation data frame}

\item{epi}{sequencing error}

\item{init}{scalar number of clusters, or vector all possible cluster numbers}

\item{prior}{prior parameters}

\item{tol}{VBEM convergence threshold}

\item{maxiter}{VBEM maximum iteration}

\item{fit_mult}{whether or not to estimate multiplicities}

\item{fit_hyper}{whether or not to estimat hyperparameters}

\item{use}{methods to get rough estimates of ccf}

\item{verbose}{show progress}
}
\value{
a list containing model parameters
}
\description{
Run Ccube with model 6: Normal-Binomial
}
