% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{CheckClonalCluster}
\alias{CheckClonalCluster}
\title{Check if the result has a clonal cluster}
\usage{
CheckClonalCluster(res)
}
\arguments{
\item{res}{Ccube result list}
}
\value{
TRUE/FALSE
}
\description{
Check if the result has a clonal cluster
}

