% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{CombineSNVandSVResults}
\alias{CombineSNVandSVResults}
\title{Combine a SNV and a SV result list.}
\usage{
CombineSNVandSVResults(resSNV, resSV)
}
\arguments{
\item{resSNV}{SNV result list}

\item{resSV}{SV result list}
}
\value{
a list containing combined results
}
\description{
Combine a SNV and a SV result list.
}
