% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cp2ap}
\alias{cp2ap}
\title{Convert ccf to vaf}
\usage{
cp2ap(x, t, cn, cr, cv, bv, epi = 0.001)
}
\arguments{
\item{x}{ccf}

\item{t}{purity}

\item{cn}{normal total copy number}

\item{cr}{total copy number in reference population}

\item{cv}{total copy number in variant population}

\item{bv}{mutation multplicity}

\item{epi}{sequencing error}
}
\value{
vaf
}
\description{
Convert ccf to vaf
}
