% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{CullEmptyCluster}
\alias{CullEmptyCluster}
\title{Remove empty clusters}
\usage{
CullEmptyCluster(model)
}
\arguments{
\item{model}{Ccube result list}
}
\value{
Ccube result list
}
\description{
Remove empty clusters
}

