% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{CullEmptyClusters}
\alias{CullEmptyClusters}
\title{Remove empty clusters}
\usage{
CullEmptyClusters(res, ssm, useEstep = T)
}
\arguments{
\item{res}{Ccube result list}

\item{ssm}{Ccub input data}
}
\value{
Ccube result list
}
\description{
Remove empty clusters
}

