% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{CullSmallClusters}
\alias{CullSmallClusters}
\title{Remove small clusters}
\usage{
CullSmallClusters(res, ssm, tol = 0.01, useEstep = T)
}
\arguments{
\item{res}{Ccube result list}

\item{ssm}{Ccub input data}
}
\value{
Ccube result list
}
\description{
Remove small clusters
}

