% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccube_R6.R
\name{GetCcf}
\alias{GetCcf}
\title{Get a rough estimate of ccf and multiplicities}
\usage{
GetCcf(mydata, use = c("use_base", "use_one"))
}
\arguments{
\item{mydata}{mutation data frame}
}
\value{
mydata mutation data frame
}
\description{
Get a rough estimate of ccf and multiplicities
}

