% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{GetCcf_sv}
\alias{GetCcf_sv}
\title{Get a rough estimate of ccf and multiplicities}
\usage{
GetCcf_sv(mydata, use = c("use_base", "use_one"))
}
\arguments{
\item{mydata}{mutation data frame}

\item{use}{methods for get rough estimates of ccf}
}
\value{
mydata mutation data frame
}
\description{
Get a rough estimate of ccf and multiplicities
}
