% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{GetPurity_sv}
\alias{GetPurity_sv}
\title{Estimate purity}
\usage{
GetPurity_sv(mydata, wgd = F, K = 6, th = 0.015)
}
\arguments{
\item{mydata}{ccube data}

\item{wgd}{whole genome duplication status}

\item{K}{total number of cluster in student't mixture}

\item{th}{threshold on weights to be eligible for clonal cluster}
}
\value{
purity
}
\description{
Estimate purity
}
