% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{HasClonalCluster}
\alias{HasClonalCluster}
\title{Check if the result has a clonal cluster}
\usage{
HasClonalCluster(res)
}
\arguments{
\item{res}{Ccube result list}
}
\value{
TRUE/FALSE
}
\description{
Check if the result has a clonal cluster
}

