% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{HasEmptyCluster}
\alias{HasEmptyCluster}
\title{Check if the result has empty cluster}
\usage{
HasEmptyCluster(res)
}
\arguments{
\item{res}{Ccube result list}
}
\value{
TRUE/FALSE
}
\description{
Check if the result has empty cluster
}

