% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp}
\alias{logsumexp}
\title{Compute log(sum(exp(x),dim)) while avoiding numerical underflow}
\usage{
logsumexp(x, margin = 1)
}
\arguments{
\item{x}{a matrix}

\item{margin}{used for apply}
}
\value{
log(sum(exp(x),dim)) a matrix of the sample size as x
}
\description{
Compute log(sum(exp(x),dim)) while avoiding numerical underflow
}

