% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MakeCcubeStdPlot}
\alias{MakeCcubeStdPlot}
\title{Make Ccube results plot}
\usage{
MakeCcubeStdPlot(ssm, res, myColors, printPlot = F, icgc = NULL,
  codeName = NULL, sampleName = NULL)
}
\arguments{
\item{ssm}{data}

\item{res}{Ccube result list}

\item{myColors}{colors}

\item{printPlot}{output flag}

\item{icgc}{icgc path}

\item{codeName}{path}

\item{sampleName}{sample name}
}
\description{
Make Ccube results plot
}

