% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{MakeCcubeStdPlot_sv}
\alias{MakeCcubeStdPlot_sv}
\title{Make Ccube results plot}
\usage{
MakeCcubeStdPlot_sv(ssm, res, myColors = gg_color_hue(10),
  printPlot = F, fn = NULL)
}
\arguments{
\item{ssm}{data}

\item{res}{Ccube result list}

\item{myColors}{colors}

\item{printPlot}{output flag}

\item{fn}{output file name}
}
\description{
Make Ccube results plot
}
