% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MapCcf2VafLinear}
\alias{MapCcf2VafLinear}
\title{Convert ccf to vaf using ccube linear mapping}
\usage{
MapCcf2VafLinear(ccf, purity, normal_cn, total_cn, mult, epi = 0.001)
}
\arguments{
\item{ccf}{ccf}

\item{purity}{purity}

\item{normal_cn}{normal total copy number}

\item{total_cn}{(average) total copy number in cancer population}

\item{mult}{(average) mutation multplicity}

\item{epi}{sequencing error}
}
\value{
vaf
}
\description{
Convert ccf to vaf using ccube linear mapping
}
