% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MapVaf2CcfLinear}
\alias{MapVaf2CcfLinear}
\title{Convert vaf to ccf using ccube linear mapping}
\usage{
MapVaf2CcfLinear(vaf, purity, normal_cn, total_cn, mult, epi = 0.001)
}
\arguments{
\item{vaf}{vaf}

\item{purity}{purity}

\item{normal_cn}{normal total copy number}

\item{total_cn}{(average) total copy number in cancer population}

\item{mult}{(average) mutation multplicity}

\item{epi}{sequencing error}
}
\value{
ccf
}
\description{
Convert vaf to ccf using ccube linear mapping
}
