% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MapVaf2CcfPyClone}
\alias{MapVaf2CcfPyClone}
\title{Convert vaf to ccf using pyclone style mapping}
\usage{
MapVaf2CcfPyClone(x, t, cn, cr, cv, bv, epi = 0.001, constraint = T,
  lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{vaf}

\item{t}{purity}

\item{cn}{normal total copy number}

\item{cr}{total copy number in reference population}

\item{cv}{total copy number in variant population}

\item{bv}{mutation multplicity}

\item{epi}{sequencing error}

\item{constraint}{whether constriant the transformation}

\item{lower}{lower bound of the constriant}

\item{upper}{upper bound of the constriant}
}
\value{
ccf
}
\description{
Convert vaf to ccf using pyclone style mapping
}

