% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ParseSnvCnaBattenberg}
\alias{ParseSnvCnaBattenberg}
\title{Parse Battenberg copy number files}
\usage{
ParseSnvCnaBattenberg(ssm, cna)
}
\arguments{
\item{ssm}{ssms}

\item{cna}{copy number}
}
\value{
ccube data frame
}
\description{
Parse Battenberg copy number files
}
