% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ParseSnvCnaConsensus}
\alias{ParseSnvCnaConsensus}
\title{Parse final consensus files}
\usage{
ParseSnvCnaConsensus(ssm, cna)
}
\arguments{
\item{ssm}{ssms}

\item{cna}{copy number}
}
\value{
ccube data frame
}
\description{
Parse final consensus files
}

