% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ParseSnvCnaOld}
\alias{ParseSnvCnaOld}
\title{Parse old cna_data.txt files}
\usage{
ParseSnvCnaOld(ssm, cna)
}
\arguments{
\item{ssm}{ssms}

\item{cna}{copy number}
}
\value{
ccube data frame
}
\description{
Parse old cna_data.txt files
}
