% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ParseSnvCnaPcawg11Format}
\alias{ParseSnvCnaPcawg11Format}
\title{Parse PCAWG-11 cna format}
\usage{
ParseSnvCnaPcawg11Format(ssm, cna)
}
\arguments{
\item{ssm}{ssms}

\item{cna}{copy number profile}
}
\value{
ccube data frame
}
\description{
Parse PCAWG-11 cna format
}

