% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ParseSnvCnaPreConsensus}
\alias{ParseSnvCnaPreConsensus}
\title{Parse new consensus files}
\usage{
ParseSnvCnaPreConsensus(ssm, cna)
}
\arguments{
\item{ssm}{ssms}

\item{cna}{copy number}
}
\value{
ccube data frame
}
\description{
Parse new consensus files
}

