% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{RemoveClusterAndReassignVariants}
\alias{RemoveClusterAndReassignVariants}
\title{Remove a (or more) cluster and reassign its data if the cluster is nonempty}
\usage{
RemoveClusterAndReassignVariants(res, removeIdx, label = NULL, ssm = NULL)
}
\arguments{
\item{res}{Ccube result list}

\item{removeIdx}{clusters to remove}

\item{ssm}{data}
}
\value{
Ccube result list
}
\description{
Remove a (or more) cluster and reassign its data if the cluster is nonempty
}

