% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{RemoveClusterAndReassignVariantsWithEMsteps}
\alias{RemoveClusterAndReassignVariantsWithEMsteps}
\title{Remove a (or more) cluster and reassign its data if the cluster is nonempty}
\usage{
RemoveClusterAndReassignVariantsWithEMsteps(res, removeIdx, ssm = NULL,
  label = NULL, tol = 1e-08, maxiter = 100, epi = 0.001,
  verbose = F, fit_mult = T, fit_hyper = T)
}
\arguments{
\item{res}{Ccube result list}

\item{removeIdx}{clusters to remove}

\item{ssm}{data}

\item{label}{assigned labels if res doesn't have label variable}

\item{tol}{stopping condition}

\item{maxiter}{maximum iteration}

\item{epi}{sequencing error}

\item{verbose}{show progress}
}
\value{
Ccube result list
}
\description{
Remove a (or more) cluster and reassign its data if the cluster is nonempty
}
