% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{RemoveClusterAndReassignVariantsWithEstep_sv}
\alias{RemoveClusterAndReassignVariantsWithEstep_sv}
\title{Remove a (or more) cluster and reassign its data if the cluster is nonempty}
\usage{
RemoveClusterAndReassignVariantsWithEstep_sv(res, removeIdx, ssm = NULL,
  label = NULL)
}
\arguments{
\item{res}{Ccube result list}

\item{removeIdx}{clusters to remove}

\item{ssm}{data}

\item{label}{assigned labels if res doesn't have label variable}
}
\value{
Ccube result list
}
\description{
Remove a (or more) cluster and reassign its data if the cluster is nonempty
}
