% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{RunCcubePipeline}
\alias{RunCcubePipeline}
\title{Run Ccube analysis}
\usage{
RunCcubePipeline(sampleName = NULL, dataFolder = NULL,
  resultFolder = NULL, runParser = F, runAnalysis = F, runQC = F,
  runAnalysisSnap = F, numOfClusterPool = NULL, numOfRepeat = NULL,
  epi = 0.001, tol = 1e-08, maxiter = 1000, multiCore = T)
}
\description{
Run Ccube analysis
}

