% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{RunPostAssignPipeline}
\alias{RunPostAssignPipeline}
\title{A pipeline to post assign events.}
\usage{
RunPostAssignPipeline(snvRes, svRes = NULL, mydata)
}
\arguments{
\item{snvRes}{SNV result list}

\item{svRes}{SV result list}

\item{mydata}{A data frame of variants to be assigned. Ideally, the data has been processed by CcubeSV model. So it should have ccube_mult1 and ccube_mult2 columns.}
}
\value{
A list containing, res, the post assigned result list and, ssm, the annotated data
}
\description{
A pipeline to post assign events.
}
