% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{SelectClusters}
\alias{SelectClusters}
\title{Select clusters with idx. Labels, responsiabilities and log-responsiabilities unchanged.}
\usage{
SelectClusters(res, idx)
}
\arguments{
\item{res}{Ccube result list}

\item{idx}{indices of the selected clusters}
}
\value{
Ccube result list
}
\description{
Select clusters with idx. Labels, responsiabilities and log-responsiabilities unchanged.
}

