% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vbsmm.R
\name{vbsmm}
\alias{vbsmm}
\title{Variational Bayesian Student-t mixture model (VB-SMM)}
\usage{
vbsmm(data, init = 2, prior, tol = 1e-20, maxiter = 1000,
  verbose = FALSE)
}
\arguments{
\item{data}{N x D data matrix}

\item{init}{k (1 x 1) or label (1 x n, 1<=label(i)<=k) or center (d x k)}

\item{prior}{prior parameters}

\item{tol}{VBEM convergence threshold}

\item{maxiter}{VBEM maximum iteration}

\item{verbose}{show progress}
}
\value{
a list containing model parameters
}
\description{
Variational Bayesian Student-t mixture model (VB-SMM)
}

