% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{WritePcawgFormats}
\alias{WritePcawgFormats}
\title{Write files in PCAWG-11 formats}
\usage{
WritePcawgFormats(ssm, res, icgc, codeName, sampleName)
}
\arguments{
\item{ssm}{data}

\item{res}{Ccube result list}

\item{icgc}{icgc path}

\item{codeName}{path}

\item{sampleName}{sample name}
}
\description{
Write files in PCAWG-11 formats
}

