% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{WritePcawgFormats_sv}
\alias{WritePcawgFormats_sv}
\title{Write files in PCAWG-11 formats, works for both CcubeCore and ccube_m6 output}
\usage{
WritePcawgFormats_sv(ssm, res, resultFolder, sampleName, allFormats = F,
  basicFormats = T, outputMult = F, outputAssignProb = F,
  outputAssign = F, outputSubStruct = F, outputCcm = F,
  outputCcmIdx = F)
}
\arguments{
\item{ssm}{data}

\item{res}{Ccube result list}

\item{resultFolder}{path to file}

\item{sampleName}{sample name}
}
\description{
Write files in PCAWG-11 formats, works for both CcubeCore and ccube_m6 output
}
