% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{CullEmptyCluster}
\alias{CullEmptyCluster}
\title{Remove empty clusters}
\usage{
CullEmptyCluster(res)
}
\arguments{
\item{res}{Ccube result list}
}
\value{
Ccube result list
}
\description{
Remove empty clusters
}

