% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{CullEmptyClusters_sv}
\alias{CullEmptyClusters_sv}
\title{Remove empty clusters}
\usage{
CullEmptyClusters_sv(res, ssm, useEstep = T, epi = 0.001)
}
\arguments{
\item{res}{Ccube result list}

\item{ssm}{Ccub input data}

\item{useEstep}{flag to use Estep for reassignment}

\item{epi}{sequencing error}
}
\value{
Ccube result list
}
\description{
Remove empty clusters
}
