% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{CullSmallClusters}
\alias{CullSmallClusters}
\title{Remove small clusters}
\usage{
CullSmallClusters(res, ssm, th = 0.01, epi = 0.001, useEstep = T)
}
\arguments{
\item{res}{Ccube result list}

\item{ssm}{Ccub input data}

\item{th}{threshold for small clusters}

\item{epi}{sequencing error}

\item{useEstep}{use Estep for reassignment}
}
\value{
Ccube result list
}
\description{
Remove small clusters
}
