% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube_sv.R
\name{CullSmallClusters_sv}
\alias{CullSmallClusters_sv}
\title{Remove small clusters}
\usage{
CullSmallClusters_sv(res, ssm, tol = 0.01, epi = 0.001, useEstep = T)
}
\arguments{
\item{res}{Ccube result list}

\item{ssm}{Ccub input data}

\item{tol}{threshold for small clusters}

\item{epi}{sequencing error}
}
\value{
Ccube result list
}
\description{
Remove small clusters
}
