% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{GenerateCopyNumberProfile}
\alias{GenerateCopyNumberProfile}
\title{Simulate a clonal copy number profile}
\usage{
GenerateCopyNumberProfile(cnPoolMaj, cnPoolMin, cnPoolMajFractions,
  cnPoolMinFractions, numVariants)
}
\arguments{
\item{cnPoolMaj}{of colors}
}
\value{
cnProfile, a matrix with the minor copy number, major copy number, and total copy number as its 1st, 2nd, and 3rd column
}
\description{
Simulate a clonal copy number profile
}
