% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{GenerateSubClonalCNProfile}
\alias{GenerateSubClonalCNProfile}
\title{Simulate a 2-subclone subclonal copy number profile.}
\usage{
GenerateSubClonalCNProfile(cnPoolMaj, cnPoolMin, cnPoolMajFractions,
  cnPoolMinFractions, numVariants, subclonal, ccfCN)
}
\arguments{
\item{cnPoolMaj}{of colors}
}
\description{
Simulate a 2-subclone subclonal copy number profile.
}
