% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ParseSnvCnaOld}
\alias{ParseSnvCnaOld}
\title{Parse old cnv_data.txt files}
\usage{
ParseSnvCnaOld(ssm, cnv)
}
\arguments{
\item{ssm}{ssms}

\item{cnv}{copy number}
}
\value{
ccube data frame
}
\description{
Parse old cnv_data.txt files
}

