% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{RunPostAssignPipeline}
\alias{RunPostAssignPipeline}
\title{A pipeline to post assign events.}
\usage{
RunPostAssignPipeline(snvRes, svRes = NULL, mydata)
}
\arguments{
\item{snvRes}{A default reference Ccube SNV or SV result list}

\item{svRes}{An optional second reference Ccube SNV or SV results list. By default the second result list is NULL. If the second list is provided, it will be combined with the first one}

\item{mydata}{A data frame of variants (SNV or SV) to be assigned. Ideally, the data has been processed by CcubeSV or Ccube model. So it should have ccube_mult1/ccube_mult2 or ccube_mult columns.}
}
\value{
A list containing, res, the post assigned result list and, ssm, the annotated data
}
\description{
A pipeline to post assign events.
}
