% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{WritePcawgFormats}
\alias{WritePcawgFormats}
\title{Write files in PCAWG-11 formats, works for both CcubeCore and ccube_m6 output}
\usage{
WritePcawgFormats(ssm, res, resultsFolder, sampleName, allFormats = F,
  basicFormats = T, outputMult = F, outputAssignProb = F,
  outputAssign = F, outputSubStruct = F, outputCcm = F,
  outputCcmIdx = F)
}
\arguments{
\item{ssm}{data}

\item{res}{Ccube result list}

\item{resultsFolder}{path to file}

\item{sampleName}{sample name}
}
\description{
Write files in PCAWG-11 formats, works for both CcubeCore and ccube_m6 output
}

