% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccube.R
\name{WritePcawgFormats}
\alias{WritePcawgFormats}
\title{Write files in PCAWG-11 formats, works for both CcubeCore and ccube_m6 output}
\usage{
WritePcawgFormats(ssm, res, resultFolder, sampleName, allFormats = F,
  basicFormats = T, outputMult = F, outputAssignProb = F,
  outputAssign = F, outputSubStruct = F, outputCcm = F,
  outputCcmIdx = F)
}
\arguments{
\item{ssm}{data}

\item{res}{Ccube result list}

\item{resultFolder}{path to file}

\item{sampleName}{sample name}

\item{allFormats}{all PCAWG output formats}

\item{basicFormats}{basic PCAWG output formats}

\item{outputMult}{format for multiplicities}

\item{outputAssignProb}{format for assignment probabilities}

\item{outputAssign}{format for assignment}

\item{outputSubStruct}{format for subclonal structure}

\item{outputCcm}{format for co-clustering matrix}

\item{outputCcmIdx}{format for variant indices in co-clustering matrix}
}
\description{
Write files in PCAWG-11 formats, works for both CcubeCore and ccube_m6 output
}
